Integrate the login page into core app
Prior to this it dealt with two separate HTML, CSS and JS packages but
now this started to become unmanageable and so it's been combined. In
order to make it work it meant merging the CSS styles, hooking up the
login component to use the input async channel, combining the HTML and
making some changes to app-state to track if the user is logged in or not.