Witan.UI now talks to localhost backend
There is a new 'api' channel for requests to be piped out to the
api (controllers/api.cljs). This uses another large multimethod. So far
the api.cljs works on the app-state itself. I'm not entirely comfortable
with this so it could change so that the actual 'what happens next'
logic is deferred to the caller by a callback.