Fixed a bug where multiple changes to permissions crashed the app
The crash was a schema violation on save. Save occurs every time an
activity finished. The schema violation was resolved by a) changing
the schema to something more permissive and b) actually adding
pending activities to the custom reset function, so they should never
actually be saved anyway. In order to calm lots of saves in quick
succession, which causes UI lag, the 1s delay before a save is now
reset when a new activity finishes (essentially a debounce).