Move page number into the app state
As is usually the case, putting it into the app state is the most useful
place for this data when it comes to manipulating it via outside
influences. There were a couple of glitches when switching between
filters which made it difficult to reset the page number. Of course,
once it's all controlled in app state it becomes trivial.