Simplify the page selection code
After a good night's sleep I was able to reduce the number of steps
required to ensure the page selection was accurate. Now, the button
changes the URL via navigation event. The nav event reads the page
number and triggers a query. The UI widget is based on off the URL.