Fix an issue where files are uploaded with no sharing information
There has been a bug on this page specifically for some time. Because
it caches, in the view, the user's groups for the purposes of building
a form with all the upload data in, if a user goes directly to the
URL but isn't logged in, when the view first loads it'll have nil data
for the user. When they then login, that data isn't updated so any
uploads end up using the nil values. Now the view will reload when a
user is logged in and will reset the form data if it looks like the
data is missing.